[music] Use in in public interface of Note & Semitones

+++ src/mc/audio/music/note/keyboard.cpp
@@ -24,7 +24,7 @@ constexpr auto isBlackKey(Note note) noexcept -> bool
         true,   // A#
         false,  // B
     };
-    return notes[static_cast<uint8_t>(note) % 12U];
+    return notes[static_cast<size_t>(static_cast<int>(note) % 12)];
 }
 
 constexpr auto isWhiteKey(Note note) noexcept -> bool { return !isBlackKey(note); }
