+++ src/mc/audio/music/note/to_hertz.cpp
@@ -10,7 +10,7 @@ namespace mc {
 template<typename T>
 constexpr auto toHertz(Note note) noexcept -> Hertz<T>
 {
-    auto ratio = pow(T{2}, static_cast<T>(static_cast<uint8_t>(note) - 69) / T{12});
+    auto ratio = pow(T{2}, static_cast<T>(static_cast<int>(note) - 69) / T{12});
     return Hertz<T>{T{440} * ratio};
 }
